Move isDirectory/isEquivalent to utils.js
