+++ utils.js
@@ -27,4 +27,40 @@ const writeFile = async (filename, content) => {
   await fs.writeFile(filename, content, 'utf8');
 };
 
-module.exports = {writeFile};
+const isDirectory = (fp) => {
+  try {
+    return fs.statSync(fp).isDirectory();
+  } catch (e) {
+    return false;
+  }
+};
+
+const isEquivalent = (a, b) => {
+  // Create arrays of property names
+  const aProps = Object.getOwnPropertyNames(a);
+  const bProps = Object.getOwnPropertyNames(b);
+
+  // If number of properties is different,
+  // objects are not equivalent
+  if (aProps.length != bProps.length) {
+    return false;
+  }
+
+  for (const propName of aProps) {
+    // If values of same property are not equal,
+    // objects are not equivalent
+    if (a[propName] !== b[propName]) {
+      return false;
+    }
+  }
+
+  // If we made it this far, objects
+  // are considered equivalent
+  return true;
+};
+
+module.exports = {
+  writeFile,
+  isDirectory,
+  isEquivalent
+};
